use numbers for user.id not bigint

+++ data/migrations/002-user.sql
@@ -1,10 +1,10 @@
 CREATE TABLE user (
-	id BIGINT PRIMARY KEY,
+	id INTEGER PRIMARY KEY,
 	slug TEXT UNIQUE,
-    email TEXT UNIQUE,
+	email TEXT UNIQUE,
 	name TEXT,
-    password TEXT,
-    active NUMBER,
+	password TEXT,
+	active NUMBER,
 	created TEXT,
 	updated TEXT
 );
